Updated Projector.cpp
Added the two Projector, one is not finished and needs information about the shader prt in the constructor.
Also added set/get for Position and Direction. It seems to be working without any include for the Vec3. However, what should we use to get the OpenGL Vec3 as it says in the document?